backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6
pandas==2.1.3
numpy==1.25.2
lxml==4.9.3
phonenumbers==8.13.25
nltk==3.8.1
scikit-learn==1.3.2
matplotlib==3.8.2
plotly==5.17.0
networkx==3.2.1
crtsh==0.1.0
dnspython==2.4.2
python-whois==0.9.4
